feat: implement asynchronous interface behind `async` feature
Asynchronous versions of the `upload` and `download` functions
are now available in an `asynchronous` module.
The `async-net` crate is used as dependency for an async UdpSocket.
