+++ Cargo.toml
@@ -10,10 +10,18 @@ readme = "README.md"
 keywords = ["tftp"]
 categories = ["network-programming"]
 
+[features]
+default = []
+async = ["dep:async-net", "dep:async-io", "dep:futures-lite"]
+
 [dependencies]
 byte-strings = "0.3"
 thiserror = "1"
 tracing = "0.1"
+
+async-net = { version = "2.0", optional = true }
+async-io = { version = "2.4", optional = true }
+futures-lite = { version = "2.6", optional = true }
 
 [dev-dependencies]
 paste = "1"
