+++ README.md
@@ -17,12 +17,12 @@ There are several TFTP crates in the rust ecosystem:
 
 All but the last only implement the server side.
 The last library seems focused on reimplementing the tftp applications,
-not so much focused on the rust library.
+not so much focused on the rust library.  
 Additionally, it is not as robust as the Python [tftpy](https://pypi.org/project/tftpy/) library.
 
 This library, `tftp-client` implements only the client as per RFC 1350,
-including the fix for the ["sorcerer's apprentice syndrome"](https://en.wikipedia.org/wiki/Sorcerer%27s_Apprentice_Syndrome).
-It is blocking-only, doesn't implement any of the additional options,
+including the fix for the ["sorcerer's apprentice syndrome"](https://en.wikipedia.org/wiki/Sorcerer%27s_Apprentice_Syndrome).  
+Currently the library doesn't implement any of the additional options,
 but provides robust control over how timeouts are handled.
 
 Unlike `rtftp`, retries include exponential backoff (with an upper limit) and
